backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
aiohttp
deepgram-sdk
groq
pyaudio
